requirements.txt
Flask
torch==2.1.0+cpu
torchvision==0.16.0
ultralytics
numpy<2.0
pillow
opencv-python

